Add property "timeofyear" to the Indexer class of module indextools.
"timeofyear" index arrays are supposed to support the handling of SeasonalParameter instances.  They map the current simulation step to the index of a simulation step in a reference year.

Note that the unittests capture the basic functionalities. But they a just adaptions of the unittests of "dayofyear" index array.

The small inline functions passed to method _calcidxs now have names, which are used for a error message. Hence, for new index array properties, also usefull names for the related functions should be selected.

Additionally, some small PEP8 related fixes were made.