Implement option "warnmissingcontrolfile".
Beforehand, a missing control file generally resulted in an exception abording the whole model initialization procedure.  Now, the user can set the option "warnmissingcontrolfile" of class Options of module "objecttools" to True. Then the mentioned exception is catched in method "initmodels" of the HydPy class, and only a warning is raised.