Implement option "checkseries".
Beforehand, one had always to make sure that the whole initialization period was completely covered by the required input data of all selected models. Otherwise, an error was raised, aborting the whole time series initialization procedure.

Usually, this is a save and useful strategy. But there are situations, where one for example wants to extend the initialization period into the future, but be able to add new input date when they become available.

To allow this, the property "checkseries" of the "Option" class of module "objecttools" has been implementend.  Additionally,  the method "load_ext" of the "IOSequence" class of module "sequencetools" has been extended.

Note that some refactoring of method "load_ext" might be advisable. And it of greatest importance to add some unit tests, as the associated indexing opperations are generally error prone!