Improve programming style.
Mainly, white lines or white spaces are added or deleted. There are also same other changes like always writing `x not in xs` instead of  `not x in xs`.

Note that these changes should not affect the functionality of the code itself.  But due to a test coverage still only around 80 %  one cannot be absolutely certain. Nevertheless, it seemed to be important to do it.  Now it should be less likely to overlook real problems marked by code inspection.