Add function `autodoc_basemodel` to module `magictools`.
The function `autodoc_basemodel` can be called in `__init__` modules of base model packages. It adds some reStructuredText lines into the module docstring, which links to the other modules and specific classes of the base model packages. This helps in building the online documentation of HydPy in a more automated manner.