Improve class `Test` of module `testtools`.
Firstly, the initialization argument `seqs` is no optional.  If it is not defined by the user, all values of the nput, flux and state sequences of the tested model as well as the values of the simulation sequences of the connected nodes are printed in the order of their mentioning.

Secondly, print also the date of each row in a user defineable format.

Thirdly, 1-dimensional sequences are now printed properly under the  premise, that their shape is (1, ). Otherwise, an exception is raised.

Much remains to be done...