Introduce the `initvalue` property of class `Parameter` and class `Sequence`.
`initvalue` returns the default parameter value or the default initial condition value definit by the class attribute `INIT`, if the `usedefaultvalues` option is set to `True`.  If this option is set to False, appropriate missing value representations are returned. If the option and set to True and not `INIT` attribute exists, the behaviour of `initvalue` differs.  For parameters, the appropriate missing value representation is returned.  For sequences, zero is returned.

This commits also changes some other methods  and doctests related to the introduction of the `initvalue` property.