Revise method `trim` of module `objecttools`.
This commits makes the `trim` method more flexible.

The first major change is that floating point and integer values are now handled in a different manner. For floating point values, the old trimming with optional warning messages is performed. For integer values, exceptions are raised instead.

The second major change is that missing values can now be handled properly.  Comparing a value with e.g. a lower boundary with a `nan` value does not result in any trimming, warnings or exceptions. For example, this offers more flexibility in defining the order in which the values of different, related parameters are set.