Handle pure Python model functions as instance methods instead of class methods.
Before this commit, model functions were added to the model class itself via a metaclass.  Now these functions are explicitely converted to methods and connected to model objects during their initialization. This strategy is more consistent with the handling of parameter and sequence objects and their respective collection classes. Due to this commit, the automatically generated documentation of application models is more consistent.