Improve class `UnitTest` of module `testtools`.
Now one is allow to start a series of examples from an arbitrary number.  For example, one can first define a test function object for examples one and two, and afterwards a different test function object for examples three and four (with the numbers 1, 2, 3 and 4 printed in the first column).