Fix method `__init__` of class `'HydPy'  of module 'hydpytools' .
Now the global information in module `pub` is only change, when a `projectname` is defined.

As a consequence, one can now initialize an additional HydPy instance during testing, without to break an already prepared project. Hence it is not necessary to restart working scripts each time a model needed to be "recythonized" and tested via an integration test.