Make parameter `KE` of HydPy-L-Land independend from time.
Beforehand, the value of `KE` was proportional to the length of the simulation time step.  This was a trick to allow for simulations with `lland_v1` with step sizes deviating from 1 day, despite the fact that it implements the Turc-Wendling evaporation equation.

Now the value of `KE` is independend of the simulation time step. To apply HydPy-L-Land e.g. on a hourly time basis, it is strongly recommended to use `lland_v2`.