Add model method tuples `_INPUT_METHODS` and `_OUTPUT_METHODS` to class `Model` of module `modeltools`.
Before `calc` methods and `update` methods were not treated seperately.  They all were included in one `_RUNMETHODS` tuple.  The method names somehow decided on when the methods were applied.  Now the distinction between the functionality of the different methods should be more clear.

On this occasion, I changed `RUNMETHODS` and `ADDMETHODS` to `RUN_METHODS` and `ADD_METHODS` respectively.  Not very important, but a slight improvement in consistency.