Improve the usage of Cython.
Modules `configutils` and `pointerutils` now have finalized classes. Also some Cython savety options are set to False.

For the cythonization of hydrological models, it is no possible to not use the `nogil` option and to set the savety options mentioned above to True. This could be helpful for debugging cythonized models.