Add the `debug_cython` option to `setup.py`.
Using this option sets same savety flags in already existling pyx files to False and circumvents using the `nogil` option and also sets the `fastcython` option of the `pub` module to False.

This option increases computation times but should result in clean error messages instead of program crashes if there is a bug in a cython extension class.