Make module `modelutils` compatible with PyInstaller.
After freezing HydPy, the command `import Cython.Build` resulted in exceptions, as some modules were missing in the executable.  This is why I moved the command to method `compile_` of class `Cythonizer`.  This is the only method which requires this import and it will (hopefully) never be used in "within" the executable.