Introduce the "solver parameters".
There is a new subparameter group called `solver`. "Solver parameters" are e.g. used for configuring numerical root finding algorithms or integration algorithms.

For completeness, a new parameter class `SolverParameter` is added.  It allows for both an automatic updating of solver values (as known from the "derived parameters"), but also to be manual value modification.