Improve the handling  `Element` and `Node` object names.
As other objects in HydPy, instances of the classes `Element` and `Node` know their own name.  Also, elements and nodes are usually accessed as attributes of instances of their collection classes `Elements` and `Nodes` (and as attributes of `Connections` instances).

To ensure that the attribute names and element names (or node names) remained consistent, it was not allowed to change the name of any element or node.  (The same is true for parameters and sequences).

This was sufficient for using already finished network structures. But it made it very hard to construct new network structures or to change existing ones.

Now some strategies are implemented which allow for changing the names of elements and nodes while keeping the whole network structure consistent.

By the way, also new `__iter__` methods are defined. Now, only the element or node is returned when iterating `Elements`, `Nodes`, or `Connections` objects instead of (name, element) or (name, node) tuples.