Generalize and improve binary comparisons of class `Variable` of module `variabletools`.
Note that variables (parameters and sequences) do now generally return a single boolean value, when they are compared with other comparable objects.  This is even the case variables handling multiple values and differs from the old behaviour. If the old behaviour is required, perform the comparisons via the `value(s)` property.