Add docstring substitution capabilities to module `autodoctools`.
Defining links using Sphinx text roles is hard to read for deeply nested package structures.  The commen Sphinx substitutions can be in conflict with Sphinx's Python roles (e.g. :class:).  This commit introduces an alternative substitution mechanism, also based on the "|something|" syntax.  This has the benefit that, when a substitution cannot be resolved, a warning is raised by Sphinx automatically.

This commit implements:
 * class `Substituter`
 * function `autodoc_applicationmode