Experiment with using `rst_prolog` for substitutions.
This is a more general solution than changing `__doc__`, as it also covers attributes like those of type `int`.  And it does not in increase import times.

However, this means that the docstrings in interactive mode and those shown in the online documentation are different.  This also affects some testing routines.