Add and correct a huge number of substitutions/links.
This commit completes commit 54325e7dc8fc7950eed4d6eb43fe250ace35eb83.

Introducing the substitutions revealed that some parts of the documentation were not up-to-date.  Primarily, `modelcollection.rst` and `development.rst` had to be updated.

For base models `dam` and `llake`, two doctests had to be adapted.  The automatic detection of relevant parameters and sequences does not work for the new substitution mechanism.