Remove `lstream_v002` ("dVdQ") and rename `lstream_v003` (external rating curve) to `lstream_v002`.
As explained in the documentation of application model `lstream_v002`, we do not see an advantage in the "dVdQ" equation when following a "hydrodynamic" approach. However, we keep the underlying methods of base model `lstream` for now.  They might prove useful for constructing a simlified, computationally more efficient application model later.