Do not install version 2.0.0 of `sphinxcontrib-bibtex` on Travis-CI.
With version 2.0.0, Sphinx needs to run twice on Travis. See issue  mcmtroffaes/sphinxcontrib-bibtex#214.

This is what I tried to get it running (the first command is allowed to fail, the second not):
  - if [ $TRAVIS_PYTHON_VERSION == "3.7" ] ; then travis-sphinx build  --source=hydpy/docs/auto || true; fi
  - if [ $TRAVIS_PYTHON_VERSION == "3.7" ] ; then travis-sphinx build  --source=hydpy/docs/auto; fi

However, the second command also fails, as Sphinx cannot find the `bibtex.json` file generated in the first run.
I suspect, this is due to how the `travis-sphinx` script works.