Add functionalities for extracting exchange item values from condition sequences (modules `itemtools`, `xmltools`, and `servertools`).
This commit restricts that `SetItem` is the only `ChangeItem` that can handle condition sequences. I don't think we need to implement the required extraction functionality for `MathItem` subclasses (until we encounter relevant use-cases).

So far, I did not spend any thought on extracting time-series values. Do we need to support it? Or should we clarify that it's not supported?