Add the land-use type `SEALED` to base model `hland` and its application models.
This commit adds the linear storage cascade to application model `hland_v3` (HBV96/PREVAH).  PREVAH does not implement such an additional runoff concentration approach, but we need it to model the time delay of the response of the sealed areas.

This commit adds, renames, and REMOVES a few derived parameters.  We now need to perform a few additional calculations within the model methods.  Still, I felt this was necessary to increase understandability and keep the overview of the different spatial extents of the sequences of the various application models.

This commit also adds a "muliple zones" integration test to each application model to make sure the water balance holds (and the related test methods work) also for realistic configurations.