Introduce piecewise linear and spline interpolation as alternatives to neural network-based interpolation.
This commit adds the new auxiliary modules `interptools` and `ppolytools` as well as their cython-complements `interputils` and `ppolyutils`.  Also, it moves class `SeasonalANN` from module `anntools` to `interptools` and renames it to `SeasonalInterpolator`.  Finally, it should adjust all parts of the documentation concerning interpolation "within" model methods to clarify that users are now free to select the interpolation algorithm of their choice.