Add class `MultiplyIUH` to module `calibtools`.
Module `calibtools` now imports module `arma_control`. We need to change this because we want to prevent importing features from the `models` subpackages into the modules defined in the `auxs` subpackage.  We should either move `ReplaceIUH` and `MultiplyIUH` to the `arma` model or define a `Responses` base class in a module somewhere in the `auxs` subpackage.