Improve the snow redistribution approach implemented in base model `hland`  and all its application models.
The simple snow redistribution approach aims to prevent the formation of "snow towers" in high elevation zones in mountainous catchments.  The old implementation works but can have the undesired side-effect of forming "snow towers" in the lowest zones of such catchments.  We knew this could theoretically happen but thought this would not happen in reality.  Experience in an alpine catchment in Switzerland showed we were wrong.

In the revised approach, we avoid this problem by a "from bottom to top" re-redistribution. The implementation is a little complicated, but the results are sensible.  To make this work, we needed to exclude all internal lake zones from participating in snow redistribution (which might be a good choice anyway).