Improve the function object `repr_` of module `objecttools`.
The original purpose of this commit is to let `repr_` return a string representation of the items of scalar numpy arrays (e.g. 3.0) instead of the original array representations (e.g. array(3.)).

I realised that the doctest module does not evaluate the related doctests because they are defined within an "attribute docstring". To satisfy both sphinx and the doctest module, I duplicated the docstring. A better solution is required!

Due to this change, a wrong doctest was reported. I fixed it.