Remove the NetCDF related options `flatten` and `timeaxis`.
From now on, reading and writing NetCDF files always works like setting `flatten` to `True` and `timeaxis` to 0. Enabling `flatten` circumvents defining partially incomplete arrays when working with sequences of different shapes.  Using the first axis for time agrees with the NetCDF-CF metadata conventions.  Also, it is the only reasonable choice for reading or writing data "just in time" during a simulation run.      We decided to fix these settings to ease the implementation of the "just in time" functionalities and other upcoming NetCDF related improvements.  Simpler usability was also a relevant point for introducing this backwards-incompatible change.