Store time series data in a single directory by default.
Until now, class `SequenceManager` was unique, as it handled three default directories: "input" for all model-specific input sequences, "output" for all model-specific output sequences, and "node" for all node sequences.  This functionality was inconsistent with the single `currentdir` inherited from class `FileManager`, and assigning values to this property had no effect.  Also, the separation between the different directories became unclear due to adding writing features to the input sequences and the increasingly ambiguous nature of the node sequences.  Hence, consistently to handling the control files, we decided to store all time series in a directory named "default" from now on.   The user can either temporarily change the directory for all sequences via property `currentdir` or define individual paths for different sequence objects via the new `dirpath` property of class `IOSequences`.

This commit also updates the XML support and addresses the issue of reading time-series data from and writing time-series data to different directories "just in time" during a SINGLE simulation run.   We leave the handling of MULTIPLE runs for a later commit.  We need to rework the related features more thoroughly due to the requirement to change directories (temporarily) from the client-side for data assimilation.