Replace the HydPy-specific NetCDF attribute `timestampleft` with `timereference` for a more precise definition of the time reference of state sequences.
The supported values of `timereference` are `left interval boundary`, `right interval boundary`, and `current time`.  The name of the related global option is still `timestampleft`, and its type is still `bool`.  However, from now on, this option only affects "interval-related" time series like precipitation or (average) temperature but not states like soil moisture.

There are still some sequences like `WaterLevel` of the `dam` model (which is a factor sequence) that are also "point-related" but handled as "interval-related".  We have to find a way to deal with these exceptional non-state sequences later.