Move application model `hstream_v1` to the new model family `HydPy-Musk`, rename it to `musk_classic`, and make its parametrisation more flexible.
This commit became much larger than expected.  We use `hstream_v1`/`musk_classic` in many doctests, which all needed to be adapted.  Furthermore, the more flexible way to define the Muskingum coefficients (based on keyword arguments) made calibration with the tools of modules `calibtools`, `itemtools`, and `xmltools` impossible.  Hence, we needed to make some progress  on how to handle the calibration of parameters via keyword arguments in general.  We added the `Parameter` class attribute `KEYWORDS` as a standardisation attempt and improved the `keywordarguments` property.  Looking at the doctests, we seem to be going in a good direction, but there are still many pitfalls left, and we still need to fill the `KEYWORDS` tuple for all relevant `Parameter` subclasses besides those of `musk_classic`.