Fix the XML support for exchanging time series via `SetItem` objects (modules `xmltools` and `itemtools`).
Previously, the preparation of the `series` attribute of the relevant `IOSequence` objects crashed because it required that the `series` attribute is already prepared.  Our tests did not detect this bug.  The `series` attribute was already prepared in all cases due to other reasons (reading or writing functionality configured under `series_io`).