Let the HydPy server support so-called "output items" (module `servertools` and `xmltools`).
Recently, we discovered that OpenDA is rather restricted when working with (multi-dimensional) exchange items used for observations.  Circumventing this limitation with our "getitem" mechanism proved difficult.  Hence, we extended the relevant parts of the usual "setitem" mechanism for factor sequences and states.  Corresponding adjustments in the OpenDA adapter split the data of the related 2-dimensional exchange items into multiple 1-dimensional exchange items.  We need to gain some experience with this stopgap.  If it works well, we will at least adjust some terms in HydPy (outputitems, OutputSequence, SetItem, GetItem...) to restore consistency.