Make class `Variable` of module `variabletools` non-generic (and improve some more or less related typing issues).
This change adds a little boilerplate code to the subclasses of `Variable`.  However, it helps to avoid recursive type definitions which Mypy cannot handle currently).  Furthermore, type narrowing of `Variable` subclasses in scripts (especially for `IOSequence`) is now much more convenient.

There a some problems with documenting the attributes `subvars`, `subpars`, and `subseqs`.  For now, we exclude them from the documentation completely.