Finish the type hints of module `devicetools`.
Typing the hidden registries could be better, but it is probably not worth the effort.  Also, the annotations of the `__new__` methods are not very precise and I need to bind the TypeVar `TypesDevices` to `Devices[Any]`.  Maybe, this is inevitable due to current Mypy limitations (missing "higher kind of types" support)?  However, it's all very low priority at the moment, but I left some ToDos in the case we must revisit these issues later.