Support "station names" with Unicode characters in NetCDF files (module `netcdftools`).
To my knowledge, NetCDF-CF restricts us from using ASCII characters, which is inconvenient when working in countries which's languages frequently use different characters.  So we decided to support the UTF-8 encoding.  Therefore, we split characters requiring more than one byte on multiple array entries.  This decision does not raise compatibility issues with other NetCDF-CF software as long as the user refrains from using non-ASCII characters.  And if the prefers to use non-ASCII characters, we can at least be confident that some other tools can cope with it (we checked this for `https://www.unidata.ucar.edu/downloads/netcdf-java/`).