Let the `nox` session  `sphinx` remove the auto-generated `.doctrees` directory.
The concrete occasion for this change is that we accidentally pushed this directory to GitHub pages, but GitHub recently started to reject it because of too large file sizes.  However, I prefer to remove not in `combine_docversions.py` but directly at the end of the `sphinx` session, as this generally decreases the risk of distributing unnecessarily storage-intensive versions of the documentation.