Support writing control files (including auxiliary files) that contain submodel definitions.
This commit moves the function `get_controlfileheader` and the method `save_controls` of class `Parameters` of module `parametertools` to class `Model` of module `modeltools` and modifies them as needed.

Also, it adds the `submodelname` attribute to class `SubmodelAdder` and lets `SubmodelAdder` instances register themselves to their master `Model` class under this name.  This connection is required to inspect a specific submodel's `SubmodelAdder` configuration.  (Perhaps, a more direct approach would be possible?).