Change how `Variable` instances react on the `len` operator and move the old functionality to the new property `numberofvalues` (module `variabletools`).
I find this change somewhat unfavourable from the typing perspective (`len` now causes a `TypeError` for some but not all variables).  However, the change seems necessary because numpy gradually makes its handling of scalar and 1-dimensional arrays with a single value stricter, which now reveals an already existing problem.  A mathematical operation (e.g. multiplication) applied on a numpy float and a 0-dimensional `Variable` instance might return a 1-dimensional, single-value array.  This only happens when the numpy flow is the first operand, which triggers numpy to analyse the `Variable` instance by calling `__len__` and `__getitem__`.  The old implementation of `__len__` causes numpy to come to false conclusions.