Complement the high water protection control of application model `sw1d_lias_sluice` with a low water protection control.
Until now, the parameters `TargetWaterlevel1` and `TargetWaterlevel2` allowed a smooth switch between "free flow " (for average conditions) and "sluice control" (for high water conditions).  Now, we replace them with the new parameters `BottomHighWaterThreshold` and `UpperHighWaterThreshold`.  And we add the parameters  `BottomLowWaterThreshold` and `UpperLowWaterThreshold` to support a similar smooth switch between "no flow" (for low water conditions) and "free flow".  Avoiding the "free flow" entirely is possible by equating `BottomLowWaterThreshold` with'BottomHighWaterThreshold` and `UpperLowWaterThreshold` with'BottomHighWaterThresholdUpper`.