Adjust the code style to Pylint 3.3.
Pylint 3.3 introduces the `too-many-positional-arguments` check., which we consider valuable.  Hence, we decided to follow it where possible (a few exceptions are necessary for some cythonized methods).  Unfortunately, this breaks the signature of a few public methods.  As those are unlikely to be used in applications with positional arguments, so we nevertheless consider this acceptable (and the effort of introducing deprecations not worth it).