Improve the performance of method `update_coefs` of class `MA` of module `armatools` by a simple "convolution approach."
The old implementation, based on a nested call of SciPy's `quad` function, was too slow in some applications.  I considered using SciPy's new `cubature` function, which is clearly an advancement over `quad` for solving double integrals, but it also did not help enough.  Maybe this will change when `cubature` allows the user to define adaptive integral bounds (there is currently a pull request for this).  The problem with `cubertare` is that it evaluates possibly expensive IUH functions too often.  Hence, I decided to implement a simple, non-adaptive "convolution approach".  Due to the equidistant node locations, it is clearly not ideal regarding convergence order.  However, it allows reusing previously calculated responses.  Besides this, it is easier to configure and results in more predictable results, which seems helpful when approximating the MA model with a smaller ARMA model.

Note that the new configuration (`nmb_nodes = 1000`) is less strict than the old one.  Currently, considering the generally envisaged accuracy, it seems acceptable, but we might want to adjust it later after first experiences in real applications.