Start import `netCDF4` "normally", not via `OptionalImport`.
`netCDF4` is listed as a standard requirement, so this change should not bother anyone.  The current occasion for this change is to prevent runtime warnings with the message "numpy.ndarray size changed, may indicate binary incompatibility.  Expected 16 from C header, got 96 from PyObject".  NumPy should suppress them, but this does not work on my computer with the latest package versions when executing doctests from PyCharm in combination with `OptionalImport` for unknown reasons.